🐛 [FIX]: highlight.controller.ts 파일명 오타 수정
